Moves lib stuff into the model directory
        Because it is all business logic!
