+++ test/store_calculations_test.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var assert      = require('assert');
+var storeStats  = require('../models/store_calculations');
+
+describe('storeStats', function() {
+  
+});
